Review fixes: no panic across FFI, C header type fix, derive Copy on vtables
- ForeignDecoder::new and ForeignSimulator::new return Option instead of
  panicking on version mismatch (panic across FFI is UB)
- C header set_seed uses uint64_t instead of size_t (32-bit safety)
- Derive Clone, Copy on vtable structs, eliminating manual field copies
- Fix Julia doc comments that said "Go"
- Fix error_ptr doc (not null-terminated)
- Remove unused thiserror dependency

+++ crates/pecos-foreign/include/pecos_foreign.h
@@ -150,7 +150,7 @@ typedef struct {
     void (*reset)(void *handle);
 
     /** Set the RNG seed for reproducibility. May be NULL. */
-    void (*set_seed)(void *handle, size_t seed);
+    void (*set_seed)(void *handle, uint64_t seed);
 
     /** Destroy the simulator and free all resources. NULL handle must be a no-op. */
     void (*destroy)(void *handle);
